Start systemd user sessions where needed.
systemd-logind never ended up getting pam_systemd.so pulled in
on logins due to it being missing in the path. We insert it
as high up as we can to avoid firing up user sessions in login
vectors that we don't want it to start. This works for console
logins where it is most needed, as this is the likely GUI user
session login we have right now.