Don't enable full hwdb for all systems.
On VMs and containers, we don't want to read the full hwdb contents
(2.9mb) since most of it is entirely cosmetical (systemctl output,
for instance). We can drop a large part for most Vms and containers. A
post-update trigger can make the full db for real hardware deployments
where the performance issue is less problematic. The full db will
then live as a runtime generated file in /etc/udev/