Fix timesyncd boottime issue, create a console subpackage
There are several units and udev rules that do expensive things that are only
useful on systems with a console; increasingly cloud VMs do not provide a console,
but only provide ssh access.

By splitting off the console units/udev rules into a separate subpackage,
we prepare for the option to not ship these units/rules in cloud setups.

Timesyncd marked itself as a "notify" type, but would in practice only
do the notification after a 100 msec wait for networking. This would hold
up a lot of the boot behind this 100 msec for no good reason, nothing
is really waiting for anything timesyncd actually provides.