systemd: Autospec creation for version 242
Enable LZ4 for coredump compression in the journal.

We've had XZ disabled due to it's unfair CPU utilization and slow
compression of journal and coredump data. It would really use a very
significant chunk of resources. This has been somewhat heated debate
due to the tradeoff which means writing lots of large core dumps.

Instead, we turn on LZ4 compression in the journal and coredumps. Lz4
is fast and small, and relatively low overhead. This will give us much
less heavy coredumps at a low cost.

I've changed liblz4.so to -Os to reduce it from 180k to 80k or so, so
this is a small impact on footprint.