Don't replace /etc/resolv.conf
An old patch to /usr/lib/tmpfiles.d/etc.conf would result in
/etc/resolv.conf *always* being replaced on boot with a symlink to
/run/systemd/resolve/resolv.conf. The intent was to work around a bug in
systemd-resolved in resolving CNAMEs recursively. This patch used the +
option in tmpfiles.d to force replacing the symlink every boot.

As that particular bug has theoretically been fixed, we'll revert to the
typical default in which /etc/resolv.conf is a symlink to
/run/systemd/resolve/stub-resolv.conf, which points to the locally
running stub resolver provided by systemd-resolved, which is the typical
configuration for systemd-based distributions.

Importantly, this change will now respect an existing /etc/resolv.conf
-- so it will not change the symlink on current Clear Linux systems, and
if the administrator replaces /etc/resolv.conf with a static file or a
symlink to another location, such as /run/systemd/resolve/resolv.conf
again, that will be preserved in subsequent reboots.